Fix override method call in constructor of ConfigurableParameters
The constructor of the ConfigurableParameters class was calling an overridable
method, getVersion, which resulted in a warning message. To fix this issue, the
getVersion method was made private.

This commit ensures that the class adheres to best practices for writing
constructors and avoids any potential issues that could arise from calling
an overridable method in a constructor.

JIRA:LIGHTY-183
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>